Update GenerateEncryptedFiles to use an existing AES key

+++ samples/GenerateEncryptedFiles.py
@@ -5,6 +5,7 @@ import string
 from dataGenerator import DataGenerator
 from azure.storage.blob import BlockBlobService, ContentSettings
 from app.aescipher import AESCipher
+from app.aeshelper import AESHelper
 from app.aeskeywrapper import AESKeyWrapper
 from app.config import Config
 config = Config("app/config.json")
@@ -22,21 +23,27 @@ wrapper = AESKeyWrapper(vault = config.azure_keyvault_url,
                         key_name = config.azure_keyvault_key_name,
                         key_version = config.azure_keyvault_key_version)
 
-# Generate new key\iv pair
-aes_key = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(32))
-aes_iv = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(16))
-keys = aes_key + aes_iv
+# Use existing aes key if it exists
+if os.path.isfile(aes_key_encrypted_filename):
+    cipher = AESHelper(config).create_aescipher_from_config()
+else:
+    # Generate new key\iv pair
+    aes_key = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(32))
+    aes_iv = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(16))
+    keys = aes_key + aes_iv
 
-print 'AES key generated. Key:' + aes_key + " IV:" + aes_iv
+    print 'AES key generated. Key:' + aes_key + " IV:" + aes_iv
 
-# Save encrypted keys
-encrypted_keys = wrapper.wrap_aes_key_local(keys, wrapper.get_public_key())
-with open(aes_key_encrypted_filename, "wb+") as aes_key_encrypted_file:
-    aes_key_encrypted_file.write(encrypted_keys)
-print 'AES key wrapped and saved to ' + aes_key_encrypted_filename
+    # Save encrypted keys
+    encrypted_keys = wrapper.wrap_aes_key_local(keys, wrapper.get_public_key())
+    with open(aes_key_encrypted_filename, "wb+") as aes_key_encrypted_file:
+        aes_key_encrypted_file.write(encrypted_keys)
 
-# Encrypt script using generated keys
-cipher = AESCipher(aes_key, aes_iv)
+    print 'AES key wrapped and saved to ' + aes_key_encrypted_filename
+
+    # Encrypt script using generated keys
+    cipher = AESCipher(aes_key, aes_iv)
+
 cipher.encrypt_file_save_file(script_filename, script_encrypted_filename)
 print script_filename + " encrypted and saved to " + script_encrypted_filename
